src/presentation/prototype/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
mysql-connector-python==8.0.30
protobuf==3.20.1
pyTelegramBotAPI==4.7.0
requests==2.28.1
urllib3==1.26.11
